backend/requirements.txt
requests
psycopg2
Flask
python-dotenv